+++ core/model/cachedbuildings.go
@@ -1,0 +1,32 @@
+// Copyright 2022 Board of Trustees of the University of Illinois.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package model
+
+import (
+	"time"
+
+	"github.com/rokwire/logging-library-go/v2/logutils"
+)
+
+const (
+	//TypeCachedBuildings type
+	TypeCachedBuildings logutils.MessageDataType = "cached buildings"
+)
+
+// CachedBuildings holds the building information from the building adapter
+type CachedBuildings struct {
+	Buildings []Building `json:"buildings" bson:"buildings"`
+	LoadDate  time.Time  `json:"load_date" bson:"load_date"`
+}