+++ CHANGELOG.md
@@ -4,7 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.10.2] - 2024-06-27
+### Fixed
+- Populate building number in wayfinding building end points [#100](https://github.com/rokwire/gateway-building-block/issues/100)
+
+### Added
+- wayfinding/floorplans end point 
+
 [2.10.1] - 2024-05-22
 ### Fixed
 - Incorrect event end times [#97](https://github.com/rokwire/gateway-building-block/issues/97)